Extract nested Profile class as separate class.
